refactor: split validated out of types index
